Moving ChangePasswordController from Users to User
